requirements.txt
# climas/requirements.txt
Flask==2.3.2
pandas==2.0.3
scikit-learn==1.3.0
plotly==5.15.0
joblib==1.3.2
numpy==1.25.2
python-dotenv==1.0.0